Make PlayerDetails screen use cubit instead of riverpod. Fix problem with not fetching tournaments when scrolled to the bottom by adding one more widget - PlayerDetailsScrollable (potentially to be removed in future)
